Çalışan İlişkilendirme Tipine Göre Kullanıcı Tipi Eklendi, Hata Altyapısı oluşturuldu, Zaten görevli kullanıcılara başka bir ilişkilendirme yapılmasının önüne geçildi, Sidebarda kullanıcının rolünün gözükmesi sağlandı

+++ src/main/java/com/team1/animalproject/service/ShelterService.java
@@ -1,17 +1,26 @@
 package com.team1.animalproject.service;
 
+import com.team1.animalproject.exception.BaseExceptionType;
+import com.team1.animalproject.exception.BusinessRuleException;
+import com.team1.animalproject.exception.ViewException;
+import com.team1.animalproject.model.Kullanici;
 import com.team1.animalproject.model.Shelter;
 import com.team1.animalproject.model.ShelterAnimal;
 import com.team1.animalproject.model.ShelterWorker;
 import com.team1.animalproject.repository.ShelterAnimalRepository;
 import com.team1.animalproject.repository.ShelterRepository;
 import com.team1.animalproject.repository.ShelterWorkerRepository;
+import com.team1.animalproject.view.utils.KullaniciTipiEnum;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
+import javax.swing.text.View;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -28,6 +37,9 @@ public class ShelterService implements IBaseService<Shelter> {
     @Qualifier("shelterAnimalRepository")
     @Autowired
     private ShelterAnimalRepository shelterAnimalRepository;
+
+    @Autowired
+    private UserService userService;
 
     @Override
     public List<Shelter> getAll() {
@@ -59,18 +71,38 @@ public class ShelterService implements IBaseService<Shelter> {
         shelterRepository.delete(t);
     }
 
-    public void addWorker(List<ShelterWorker> shelterWorkers) {
-        shelterWorkerRepository.save(shelterWorkers);
-    }
-
     public List<ShelterWorker> getWorkersByShelterId(String shelterId) {
         return shelterWorkerRepository.findByShelterId(shelterId);
     }
 
     @Transactional
     public void saveWorker(List<ShelterWorker> shelterWorkers, String shelterId) {
-        shelterWorkerRepository.deleteByShelterId(shelterId);
-        shelterWorkerRepository.save(shelterWorkers);
+        boolean gorevli = shelterWorkers.stream().anyMatch(shelterWorker -> userService.kullaniciBaskaYerdeGorevliMi(shelterWorker.getWorkerId(), KullaniciTipiEnum.SHELTER));
+        if(!gorevli){
+            List<ShelterWorker> all = shelterWorkerRepository.findAll();
+
+            all.stream().forEach(shelterWorker -> {
+                Kullanici kullanici = userService.findById(shelterWorker.getId()).get();
+                kullanici.setKullaniciTipi(KullaniciTipiEnum.NORMAL_USER.getId());
+                userService.save(kullanici);
+            });
+
+            shelterWorkerRepository.deleteByShelterId(shelterId);
+            shelterWorkerRepository.save(shelterWorkers);
+
+            shelterWorkers.stream().forEach(shelterWorker -> {
+                Kullanici kullanici = userService.findById(shelterWorker.getWorkerId()).get();
+                kullanici.setKullaniciTipi(KullaniciTipiEnum.SHELTER.getId());
+                userService.update(kullanici);
+            });
+
+        }else {
+            FacesContext context = FacesContext.getCurrentInstance();
+            context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_FATAL, BaseExceptionType.KULLANICI_ZATEN_GOREVLI.getValidationMessage(), null));
+            context.getExternalContext().getFlash().setKeepMessages(true);
+            throw new ViewException();
+        }
+
     }
 
     public List<ShelterAnimal> getAnimalsByShelterId(String shelterId) {
